condense most contract code into one file
